requirements.txt
fastapi
uvicorn
google-genai
pydantic
firebase-admin
resend